// app/src/generated/models/Csa_projectnotesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Csa_projectnotesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Csa_projectnotesstatecode = keyof typeof Csa_projectnotesstatecode;
export const Csa_projectnotesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Csa_projectnotesstatuscode = keyof typeof Csa_projectnotesstatuscode;

export interface Csa_projectnotesBase {
  csa_name?: string;
  "csa_ProjectId@odata.bind": string;
  csa_projectnoteid: string;
  csa_text: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Csa_projectnotesstatecode;
  statuscode?: Csa_projectnotesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Csa_projectnotes extends Csa_projectnotesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  csa_projectidname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  csa_projectid?: object;
  _csa_projectid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}